+++ pyweather.py
@@ -1354,12 +1354,12 @@ while True:
 
     elif moreoptions == "1":
         # Or condition will sort out 3 potential conditions.
-        logger.debug("alertsPrefetched: %s ; alerts cache time: %s" % 
-                         (alertsPrefetched, time.time() - cachetime_alerts))
-        # If alerts aren't fetched at boot, there is no cache time available. Make a psuedo cache time to make it work here.
-        # Better solution coming never.
-        if alertsPrefetched == False:
-            cachetime_alerts = time.time()
+        try:
+            logger.debug("alertsPrefetched: %s ; alerts cache time: %s" %
+                             (alertsPrefetched, time.time() - cachetime_alerts))
+        except:
+            logger.debug("alertsPrefetched: %s" % alertsPrefetched)
+
         if (alertsPrefetched is False or time.time() - cachetime_alerts >= cache_alertstime and cache_enabled == True
             or refresh_alertsflagged == True):
             print(Fore.RED + "Alerts wasn't prefetched, the cache expired, or alerts was flagged", 
@@ -1689,11 +1689,6 @@ while True:
         totaldetailedHourly10Iterations = 0
         logger.debug("tenday_prefetched: %s ; refresh_hourly10flagged: %s" %
                     (tenday_prefetched, refresh_hourly10flagged))
-        # We try to display the hourly 10 cache time, if the user has fetch 10 day hourly on startup enabled.
-        # If the user does not have it displayed, display nothing but a slightly informative message.
-        # Also create a psuedo cache time to make the if statement below work.
-        if tenday_prefetched == False:
-            cachetime_hourly10 = time.time()
         try:
             logger.debug("hourly 10 cache time: %s" % time.time() - cachetime_hourly10)
         except:
@@ -2886,9 +2881,6 @@ while True:
     elif moreoptions == "7":
         logger.info("Selected option: almanac")
         print(Fore.RED + "Loading...")
-        # Create a psuedo cache time if almanac data isn't prefetched.
-        if almanac_prefetched == False:
-            cachetime_almanac = time.time()
         try:
             logger.debug("almanac_prefetched: %s ; almanac cache time: %s" %
                          (almanac_prefetched, time.time() - cachetime_almanac))
@@ -2969,9 +2961,6 @@ while True:
     elif moreoptions == "10":
         print(Fore.RED + "Loading...")
         print("")
-        # Create a psuedo cache time if sundata hasn't been prefetched.
-        if sundata_prefetched == False:
-            cachetime_sundata = time.time()
 
         try:
             logger.debug("sundata_prefetched: %s ; sundata cache time: %s" %