+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Fixed a bug where wind chill data in celsius in historical hourly displayed "kph" instead of a degree C symbol.
 * Fixed a bug where the 10 day forecast cache expire time was running off of the hourly cache time.
 * Fixed a bug where PyWeather wouldn't properly catch no moonrise data, resulting in a crash.
+* Fixed a bug where if alerts data wasn't getting prefetched at boot, when detailed alerts data was called up PyWeather would crash.
 * Fixed a potential bug where proper code to catch invalid sunrise/sunset data was not implemented.
 
 **OTHER CHANGES**